chore: 패키지 도메인 단위로 분리
