+++ src/main/java/kitchenpos/domain/menu/MenuProducts.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
+import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.exception.menuProductException.InvalidMenuProductsPriceException;
 
 import javax.persistence.CascadeType;
