+++ src/main/java/kitchenpos/domain/menuGroup/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menuGroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
