+++ src/main/java/kitchenpos/domain/table/OrderTable.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
+import kitchenpos.domain.tableGroup.TableGroup;
 import kitchenpos.exception.orderTableException.IllegalOrderTableGuestNumberException;
 
 import javax.persistence.Entity;
