+++ src/main/java/kitchenpos/dto/response/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.response;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menuGroup.MenuGroup;
 
 public class MenuGroupResponse {
