+++ src/main/java/kitchenpos/dto/response/ProductResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto.response;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 import java.math.BigDecimal;
