+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.application.menuGroup.MenuGroupService;
+import kitchenpos.domain.menuGroup.MenuGroup;
 import kitchenpos.dto.request.MenuGroupRequest;
 import kitchenpos.dto.response.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
