+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.application;
 
+import kitchenpos.application.product.ProductService;
 import kitchenpos.dto.request.ProductRequest;
 import kitchenpos.dto.response.ProductResponse;
 import kitchenpos.exception.InvalidPriceException;
