+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -1,5 +1,11 @@
 package kitchenpos.domain;
 
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItems;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.table.OrderTable;
 import kitchenpos.exception.orderException.IllegalOrderStatusException;
 import org.junit.jupiter.api.Test;
